implements print command fully.
The main idea is to start from pCursor where find command left off, and traverse the trie looking for the next word. Once we find it, we set pCursor to that, and leave (findNextWord). Then, we print the word by following parent pointers starting from pCursor up to the root, utilizing the tail node child pointers which point to parent (printWord). Finally, we prepare for the next command by shifting pCursor one unit, first by trying child, if not then sibling, if not then parent thru tail node child pointer. If we end up reaching pSuffix, then we know there are no more words to print.